// src/app/models/order.model.ts
export interface Order {
    TST_LLAVE: number;
    TST_CODIGO_FISICO: string;
    TST_TIPO: string;
    TST_FECHA: string;
    TST_EMPRESA?: string;
    TST_CLIENTE: string;
    CMA_NOMBRE?: string;
    TST_DIRECCION: string;
    TST_SUBTOTAL: number;
    TST_IVA: number;
    TST_TOTAL: number;
    TST_GARANTIA: string;
    TST_PERSONA_CONTACTO?: string;
    TST_FECHA_RECEPCION: string;
    TST_CAPACITACION?: string;
    TST_OBSERVACION?: string;
    TST_USUARIO_CREA: string;
    TST_ESTADO: string;
  }
  